requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
anthropic==0.34.0
aiosqlite==0.20.0
httpx==0.27.2
python-multipart==0.0.9
pdfplumber==0.11.4
pymupdf==1.24.10
pillow==10.4.0
pytesseract==0.3.13
